requirements.txt
discord
python-dotenv
requests
polls
